update: sua ten trang

+++ admin/dashboard/nhanvien/xoa_process.php
@@ -1,8 +1,16 @@
 <?php 
-	require  '../../connect.php' ;
-	$ma = $_GET['ma'];
+session_start();
+require  '../../connect.php' ;
+$ma = $_GET['ma'];
+
+
+if ($ma == $_SESSION['ma_nv']) {
+	echo "Bạn không thể tự xóa chính mình!";
+}else{
 	$sql = "delete from nhan_vien where ma = $ma";
 	mysqli_query($connect, $sql);
-			// die($sql);
+					// die($sql);
 	mysqli_close($connect);
 	header("Location: index.php");
+	echo 'hi';
+}
